# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anony-mail API",
    "version": "0.3.0",
    "summary": "REST + Server-Sent Events API for an inbound-only disposable email backend.",
...
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "contact": {
      "name": "anony-mail source & issues",
      "url": "https://github.com/1mrnewton/anony-mail"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development (default API_BIND_ADDR)"
    },
    {
      "url": "https://{host}",
      "description": "Deployed instance",
      "variables": {
        "host": {
          "default": "api.example.com",
          "description": "Hostname where your anony-mail instance is reachable."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/addresses": {
      "post": {
        "operationId": "createAddress",
        "summary": "Create a disposable address",
        "description": "Creates a new mailbox and returns it **together with its one-time `owner_token`** — store the token client-side; it is required for extend, rotate, and all delete operations and is never returned again. The request body is **optional**:\n\n- Omit the body entirely (or send `{}`) to get a random 10-character local part on the first configured domain.\n- Provide `domain` to choose among the configured domains.\n- Provide `local_part` to request a custom address. It must be 1–64 characters of `[a-z0-9._-]` and must not start or end with a separator (`.`, `_`, `-`). If the address is already taken, the request fails with `409 Conflict`.\n\nReserved role/validation local parts (`admin`, `postmaster`, `webmaster`, `abuse`, `ssladmin`, ... — RFC 2142 plus the CA/Browser-Forum set, extensible server-side via `RESERVED_LOCAL_PARTS`) cannot be claimed and fail with `400 Bad Request`.\n\nCreation is rate limited per IP (including a daily quota); exceeding it fails with `429 Too Many Requests`.\n\nOn servers with custom domains enabled, `domain` may also be a **verified custom domain** — send its `amd_…` claim token as `Authorization: Bearer` (missing/wrong token fails with `401`, an unverified domain with `400`).\n\nThe new mailbox expires after the server's default TTL (`DEFAULT_TTL_SECONDS`, 1 hour by default).\n\nOn servers with `ENTITLEMENTS_ENFORCED=true`, tier policy applies: custom local parts, domains beyond the free `domain_count`, and custom domains may require pro. Send the client token from `POST /api/entitlements/verify` as `X-Client-Token`; without one the request is treated as free tier.\n\nWhen the token is **attested** (App Attest, `POST /api/client/attest`), the tier's `active_mailboxes` cap is enforced per device: once the device holds that many live mailboxes, creation fails with `403` / `code: mailbox_cap` until one expires or is deleted.",
        "tags": [
          "Mailboxes"
        ],
        "requestBody": {
          "required": false,
          "description": "Optional address creation options.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressRequest"
              },
              "examples": {
                "random": {
                  "summary": "Random address on the default domain (empty body)",
                  "value": {}
                },
                "domainOnly": {
                  "summary": "Random local part on a chosen domain",
                  "value": {
                    "domain": "example.com"
                  }
                },
                "custom": {
                  "summary": "Custom local part on a chosen domain",
                  "value": {
                    "local_part": "jane.doe",
                    "domain": "example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mailbox created. The `owner_token` appears **only** in this response (and in a rotate response); it cannot be recovered later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAddressResponse"
                },
                "examples": {
                  "random": {
                    "summary": "Randomly generated address",
                    "value": {
                      "address": "a1b2c3d4e5@example.com",
                      "domain": "example.com",
                      "created_at": "2026-07-01T11:23:45.123456Z",
                      "expires_at": "2026-07-01T12:23:45.123456Z",
                      "owner_token": "am_hK3T9qmZ2xWfJ8dLpR5vY0cN1aQ6sB4eU7gIoPkXmVs"
                    }
                  },
                  "custom": {
                    "summary": "Custom address",
                    "value": {
                      "address": "jane.doe@example.com",
                      "domain": "example.com",
                      "created_at": "2026-07-01T11:23:45.123456Z",
                      "expires_at": "2026-07-01T12:23:45.123456Z",
                      "owner_token": "am_hK3T9qmZ2xWfJ8dLpR5vY0cN1aQ6sB4eU7gIoPkXmVs"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "The requested domain is a custom domain and the claim token is missing or wrong — or the supplied `X-Client-Token` is expired/invalid (`code` says which; refresh it via `POST /api/entitlements/verify`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "The expected authentication scheme.",
                "schema": {
                  "type": "string",
                  "examples": [
                    "Bearer"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "custom domains require their claim token"
                }
              }
            }
          },
          "403": {
            "description": "Rejected by tier policy (`ENTITLEMENTS_ENFORCED`). The `code` field says which gate: `pro_required` for custom local parts, extra domains, or custom domains on an insufficient tier; `mailbox_cap` when the attested device already holds the tier's maximum number of active mailboxes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "custom local parts require a higher tier",
                  "code": "pro_required"
                }
              }
            }
          },
          "409": {
            "description": "The requested custom address already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "address already exists: jane.doe@example.com"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        }
      ],
      "get": {
        "operationId": "getMailbox",
        "summary": "Get mailbox info / check existence",
        "description": "Returns the mailbox's metadata if it exists. Note that an expired-but-not-yet-purged mailbox is still returned here; compare `expires_at` against the current time to determine whether it is still active.",
        "tags": [
          "Mailboxes"
        ],
        "responses": {
          "200": {
            "description": "The mailbox exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                },
                "example": {
                  "address": "a1b2c3d4e5@example.com",
                  "domain": "example.com",
                  "created_at": "2026-07-01T11:23:45.123456Z",
                  "expires_at": "2026-07-01T12:23:45.123456Z"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteMailbox",
        "summary": "Delete a mailbox",
        "description": "Deletes the mailbox and all of its messages and attachments (cascading). Requires the owner token. Returns `204 No Content` on success.",
        "tags": [
          "Mailboxes"
        ],
        "security": [
          {
            "ownerToken": []
          }
        ],
        "responses": {
          "204": {
            "description": "Mailbox deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/extend": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        }
      ],
      "post": {
        "operationId": "extendMailbox",
        "summary": "Extend a mailbox's expiry",
        "description": "Pushes the mailbox's `expires_at` to now + the server default TTL (`DEFAULT_TTL_SECONDS`). The extension is relative to the current time, not to the previous expiry. Requires the owner token (which stays the same — extending never rotates it). No request body is required.\n\nOn servers with `ENTITLEMENTS_ENFORCED=true`, the mailbox's total lifetime (`expires_at - created_at`, extends included) is capped per tier (`max_lifetime_seconds` in `GET /api/capabilities`). Extending past the cap returns `403` with `code: lifetime_cap`; send a pro `X-Client-Token` for the higher ceiling.",
        "tags": [
          "Mailboxes"
        ],
        "security": [
          {
            "ownerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Expiry extended; the updated mailbox is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                },
                "example": {
                  "address": "a1b2c3d4e5@example.com",
                  "domain": "example.com",
                  "created_at": "2026-07-01T11:23:45.123456Z",
                  "expires_at": "2026-07-01T13:05:10.900000Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "Extending would exceed this tier's maximum mailbox lifetime (`code: lifetime_cap`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "extending would exceed this tier's max mailbox lifetime of 86400s",
                  "code": "lifetime_cap"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/rotate": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        }
      ],
      "post": {
        "operationId": "rotateOwnerToken",
        "summary": "Rotate the owner token",
        "description": "Reissues the mailbox's owner token and returns the new one; the previous token is revoked **immediately**. Use this if the token may have leaked. Requires the current owner token. No request body is required.",
        "tags": [
          "Mailboxes"
        ],
        "security": [
          {
            "ownerToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Token rotated. This response is the only place the new token appears.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotateTokenResponse"
                },
                "example": {
                  "owner_token": "am_Zx8cV2bN4mA6sD1fG3hJ5kL7qW9eR0tY_uIoPaSdFgs"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/MailboxNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/addresses/{address}/messages/{id}/read": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        }
      ],
      "post": {
        "operationId": "markMessageRead",
        "summary": "Mark a message as read",
        "description": "Sets the message's `seen` flag. Idempotent — marking an already-read message succeeds. Open like the read endpoints (knowing the address already grants read access), so no owner token is required.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "204": {
            "description": "Message marked read."
          },
          "404": {
            "$ref": "#/components/responses/MessageNotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/addresses/{address}/messages/{id}/raw": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AddressPath"
        },
        {
          "$ref": "#/components/parameters/MessageIdPath"
        }
      ],
      "get": {
        "operationId": "getRawMessage",
        "summary": "Download the original message (.eml)",
        "description": "Returns the message's original RFC 5322 bytes as received over SMTP (dot-unstuffed), served as `message/rfc822` with an attachment disposition (`{id}.eml`) and `X-Content-Type-Options: nosniff`. Only available when the server runs with `STORE_RAW_MESSAGE=true`; otherwise — and for messages delivered before the toggle was enabled — this endpoint returns `404`.",
        "tags": [
          "Messages"
        ],
        "responses": {
          "200": {
            "description": "The original message bytes.",
            "headers": {
              "Content-Disposition": {
                "description": "Attachment disposition named after the message id.",
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=\"550e8400-e29b-41d4-a716-446655440000.eml\""
              },
              "X-Content-Type-Options": {
                "description": "Always `nosniff`.",
                "schema": {
                  "type": "string",
                  "enum": [
                    "nosniff"
                  ]
                }
              }
            },
            "content": {
              "message/rfc822": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "description": "Raw RFC 5322 message bytes."
                }
              }
            }
          },
          "404": {
            "description": "Message not found, or raw retention is disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "raw message not available (missing, or raw retention is disabled)"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ownerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-mailbox owner token returned once by `POST /api/addresses` (and by rotate). Format: `am_` + 43 base64url characters. Sent as `Authorization: Bearer <owner_token>`."
      },
      "claimToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-domain claim token returned once by `POST /api/custom-domains`. Format: `amd_` + 43 base64url characters. Sent as `Authorization: Bearer <claim_token>`."
      },
      "clientToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Token",
        "description": "Signed tier token (12h TTL) from `POST /api/entitlements/verify` (unattested) or `POST /api/client/attest` / `assert` (attested, App Attest). Optional: on enforcing servers a request without it is treated as free tier — unless the server reports `attestation.required`, in which case mutations demand an attested one. `Authorization` stays reserved for owner/claim tokens, so both can be sent together."
      }
    },
...
    "responses": {
...
      "TooManyRequests": {
        "description": "A per-IP rate limit or quota was exceeded. Retry later.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "daily address creation limit reached"
            }
          }
        }
      },
...
    },
...
  }
}